--- src/ai/backend/manager/services/group/BUILD
@@ -1,1 +1,0 @@
-python_sources(name="src")
