--- src/ai/backend/manager/services/group/base.py
@@ -1,18 +1,0 @@
-from dataclasses import dataclass
-from typing import override
-
-from ai.backend.manager.actions.action import BaseAction, BaseBatchAction
-
-
-@dataclass
-class GroupAction(BaseAction):
-    @override
-    def entity_type(self):
-        return "group"
-
-
-@dataclass
-class GroupBatchAction(BaseBatchAction):
-    @override
-    def entity_type(self):
-        return "group"
