+++ src/ai/backend/manager/services/groups/actions/usage_per_month.py
@@ -1,25 +1,30 @@
 from dataclasses import dataclass
 from typing import Any, Optional, override
+from uuid import UUID
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.user.base import UserAction
+from ai.backend.manager.services.groups.actions.base import GroupAction
 
 
-# TODO: Batch action으로 만든 후 entity_id엔 모든 user_id를 넣어야 함.
+# TODO: Batch action 변경
 @dataclass
-class AdminMonthStatsAction(UserAction):
+class UsagePerMonthAction(GroupAction):
+    month: str
+    group_ids: Optional[list[UUID]] = None
+
     @override
     def entity_id(self) -> Optional[str]:
         return None
 
     @override
     def operation_type(self):
-        return "admin_month_stats"
+        return "usage_per_month"
 
 
 @dataclass
-class AdminMonthStatsActionResult(BaseActionResult):
-    stats: list[Any]
+class UsagePerMonthActionResult(BaseActionResult):
+    # TODO: 리턴 타입 만들 것.
+    result: list[Any]
 
     @override
     def entity_id(self) -> Optional[str]:
