--- src/ai/backend/manager/services/user/BUILD
@@ -1,1 +1,0 @@
-python_sources(name="src")
