--- src/ai/backend/manager/services/user/base.py
@@ -1,18 +1,0 @@
-from dataclasses import dataclass
-from typing import override
-
-from ai.backend.manager.actions.action import BaseAction, BaseBatchAction
-
-
-@dataclass
-class UserAction(BaseAction):
-    @override
-    def entity_type(self):
-        return "user"
-
-
-@dataclass
-class UserBatchAction(BaseBatchAction):
-    @override
-    def entity_type(self):
-        return "user"
